Fully configure continuous integration (#22)
Configure CI to clone the oscar-exercises repository to test all the instructor notebooks.

This makes sure that any changes made to OSCAR don't break the existing exercises.  The exercises repo remains separate from the oscar repo for ease of installation, and for separation of the exercises from the source code (since we don't always want all the exercise solutions to be present in the same location as the source, e.g. when we install on student machines).